requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
et_xmlfile==2.0.0
fastapi==0.128.0
h11==0.16.0
idna==3.11
mercadopago==2.3.0
openpyxl==3.1.5
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
requests==2.32.5
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
yt-dlp==2026.3.3
supabase==2.3.4
